requirements.txt
gevent
urllib3
chardet
requests
humanize
asyncpraw
dateparser
youtube_dl
translators
python-dotenv
discord-together

git+https://github.com/Pycord-Development/pycord